Add documentation describing the process to propose new GT4Py features. (#18)
An attempt to formalize the process to propose and discuss relevant new features in GT4Py, based on standard practices in other well-known libraries in the Python ecosystem like NumPy. Later PRs could improve the discovery and reading of GDPs, by providing automatic synchronization of the docs/GDPs folder with the repository wiki.