Refactored frontend parsing of `with` blocks (#124)
Currently each `with computation(...), interval(...): ...` was canonicalized by rewriting the AST into nested `with` blocks of the form
```
with computation(...):
  with interval(...):
```
and a later two-stage processing into a single `ComputationBlock` DIR node. This is not only hard to maintain and deviates from the DIR, which should be close to the actual gtscript definition, but also hard to expand. In preperation of the addition of the location/cell type specification for non-cartesian grids this part was therefore refactored. Now each `with` block is parsed in a single method and `with computation(...): with interval(...)` blocks are rewritten into `with computation(...), interval(...): ...`.

* Refactored frontend parsing of `With` blocks
* Added reordering of nested `with` statements (including tests)

Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>