Fix VarRef data type bug in NumPySourceGenerator._visit_branch_stmt (#138)
This PR resolves an issue in the numpy backend that occurs when the target of an if branch does not have a data_type member. This is true for FieldRef but not VarRef objects. The solution is to fetch the data type for each VarRef from the corresponding VarDecl in the block_info.symbols table.