Use unique arg index (#140)
This PR fixes the generation of unique arg indices for fields. Strictly speaking, it is only needed to have two unique index sets, one for temporary fields and another for non-temporary fields (normal fields and global parameters). This PR generates only one set of indices, so every index will be truly unique in the source file.