Cumulative performance counters (#245)
Make each stencil object capable of populating the exec_info dict with a subdict of diagnostics. These include:

    the number of calls;
    the call and run time of the last call;
    the total call and run time (over all stencil invocations).

The key under which the subdict is added coincides with the stencil class name. To ensure back-compatibility, this feature is disabled by default, and enabled only if exec_info["__aggregate_data"] is True.