Add resolved_values_list to func_externals (#279)
This ensures that the externals imported in the resolved functions are added to the stencil fingerprint.

An earlier PR worked for functions that were themselves externals. I recently found out that imported externals in functions that are resolved implicitly (no explicit from __externals__ import func) are not included in the resolved externals (and hence the fingerprint of the stencil). This is a clear bug.