Enhance DemoteLocalTemporariesToVariablesPass (#276)
This PR is intended to expand the number of temporary fields that can be reduced to scalars by the DemoteLocalTemporariesToVariablesPass. The existing implementation will not demote any temporary that is referenced across more than one stage. However, a temporary field can also be demoted to a scalar if it is always written before being read in each stage where it appears.