[Field dimensions] Support field masking (#288)
Co-authored by: @jdahm 
Co-authored by: @eddie-c-davis

This PR is the first in a series of PRs that addresses lower dimensional storages. This PR primarily allows for lower-dimensional storages to be used in stencils, both in the annotations and in execution with storages constructed using the mask argument. This:

    Adds frontend checks for indexing lower-dimensional fields
    Ensures the offsets are added only for dimensions that exist in the field
    Fixes debug, numpy, and GT backends to support lower-dimensional fields
    Stencil object checks support storages using mask
    Frontend and backend tests

A follow-up PR will add back the storage reduction pass.