Improve test suite readability (#392)
Cleanups in test suites.

This commit modifies the _run_test_implementation method of the test suites. It removes a lot of unnecessary code, adds some comments and improves readability in general. Additionally, auxiliar nametuples classes like SymbolKind and _SymbolStrategy are replaced by dataclasses, and domain axes are now supported as part of the field specification.

Co-authored-by: Enrique G. Paredes <enriqueg@cscs.ch>