Do not visit empty computation blocks in `DefIRToGTIR` (#415)
This PR prevents a `non-empty vertical loop` error from occurring in the `GreedyMerging` pass when a vertical loop section has no executable horizontal sections. This can occur when an inlined if wraps all code in an interval and evaluates to false at compile time. For example,

```python
def stencil(q: Float3D, a: Float3D):
    from __externals__ import iv
    ...
    with computation(FORWARD), interval(1, 2):
        if __INLINED(iv == -2):
            q = (3.0 * (a[0, 0, -1] + a) - q[0, 0, -1])
...   
stencil(q, a, externals={"iv": -1})
...
```
The proposed solution is not to visit the empty blocks in `DefIRToGTIR` so that the empty horizontal executions are never created in the GTIR.