Preserve API order for module params in GTIR (#412)
This PR modifies the make_args_data_from_gtir method in the ModuleGenerator class to use lists rather than sets for the referenced_field_params and referenced_scalar_params variables to preserve the API signature order in field_info and parameter_info. This is necessary for the current stencil wrapper in fv3core FV3 implementation.