Automatically synchronizing GPU storage (#409)
This PR modifies the GPUStorage (managed memory) class to keep track of modified storages in a class level dictionary. When a storage contained in the dictionary is accessed outside of a stencil (e.g,. in __getitem__), the GPU is synchronized and the dictionary will be cleared. Note that this feature is intended to be used with stencils compiled with the device_sync backend option set to False, otherwise the device will be excessively synchronized.