Mask statement inlining OIR pass (#416)
Mask statements in OIR are currently implemented as boolean fields that replace the logical expressions in `FieldIf` statements. This can result in an excessive number of temporary fields in codes with many conditionals (microphysics for example). This PR implements a `MaskInlining` OIR pass that replaces the boolean fields with the actual logical expressions (e.g., `FieldAccess` -> `BinaryOp`) to reduce the number of temporary fields.