GDP-3: Duck Storages [declined] (#28)
This GDP and its reference implementation have been declined after several discussions.

The main points why this GDP is not the right way for storages are as follows.

The original motivation for this GDP was that moving away from the design where storages are subclasses of np.ndarray. The GDP proposed to make this change so that one-sided changes to coupled host/device buffers managed by GT4Py storages can be more reliably tracked, since functionality is not handled by the parent class. This issue is indeed solved by the GDP. However, it just moves the problem to a new one: The changes that can now be tracked will often require the allocation of a new buffer, for which performance optimal allocation can not be guaranteed based on the information available.

Resolving this in any way leads to a storage that is either severely limited in functionality, which prompts users to escape to different frameworks anyways (as is the case with the old storages), or that performance may be severely impacted depending on the exact use of the storages, which is not transparent to the user. In other cases the behavior may be unintuitive (broadcasting, slicing). Overall it seems impossible to find a solution that is a good trade off for a substantial share of all use cases.

Instead, we currently propose to pursue a "No Storage" approach in the future. The idea there would be that gt4py provides utilities that allocate buffers which have desireable properties for stencils, yet the framework will not provide any methods on the buffers, i.e. it is up to the user to fill values in the allocated memory, implement ufuncs if desired etc.

However, a generic interface like the __gt_data_interface__ proposed in this GDP is still assumed to be necessary.