functional: Tuple support in embedded mode (#629)
Add tuple support in embedded iterator view.

Includes fix bad design in the original field view design: stencil closure should have only one output symbol (which can refer to a tuple of fields).

Implemented semantics:
For output fields, fields of tuples or tuples of fields both work for an applied stencil that returns tuple.
For input fields, fields of tuples or tuples of fields are both treated as iterator of tuples. If you need separate iterators, pass the fields as separate arguments, not as tuple.

See ADRs for details.