Add gtc:dace:gpu backend (#731)
This PR adds the gtc:dace:gpu backend and renames gtc:dace to gtc:dace:cpu.

This is done based on the soon-to-be-outdated expansion that is currently in master. The intention is to already adapt the infrastructure for a dace GPU backend such that the backend refactor can be meaningfully be broken down to more digestable chunks.