Fix bugs when using __future__ annotations - PEP 563 (#744)
PEP 563 delays the evaluation of type hints after the definition time. Type hints are converted into string literals and stored in __annotations__. This PR proposes to evaluate argument annotations of the definition function which are strings unless they are present in the dtypes dictionary (so to still accommodate for parametrized data types).