Broadcasting support (#797)
Added support for broadcasting, namely a `broadcast` operator which enables broadcasting of fields.

The syntax is as follows: `broadcast(a, (A, B))` where `a` is a `LocatedField` and `A, B` are of type `Dimension`.

During lowering to iterator IR the broadcasted dimensions are disregarded. **Note:** Executing a `shift` in a dimension of a broadcasted `Field`, on which the non-broadcasted `Field` was not defined is supported.