Always include args and loop over param instead of api_signature (#843)
Params was not updated after pruning arguments, which left fields that were only referenced in eliminated horizontal regions still in the signature in the stencil module, but not in the pybind11 generated run method and caused a mismatch. This adds a test and fixes the issue.

It turns out that the easy fix is to not prune parameters at all, but to loop over the node.params instead of the api_signature when creating the internal bindings.