Update Jfrog docker registry URL (#951)
Update JFrog artifactory URL. See also #953 

Somehow the artifactory is not available anymore at the URL pointed to by the CSCS-CI organization's CSCS_REGISTRY variable. This breaks all CIGR pipelines in general.
The existing credentials work however on the new location. The reason the new URL is hardcoded into the gitlab-ci.yml is that we can not change the externally managed variable.