Fix return type of fieldoperator (#1027)
While not explicitly written in the ADR #956, I think this is the behavior that we discussed, otherwise it's not possible to chain field_operators that return scalars.

As a side-effect, it enables using field_operators from scans. We can discuss if we want to change this behavior longer term.

Additional: fix the gtfn code generator to work with scans without argument.