Remove ellipsis as (generic) field dimension  (#1106)
Very early in the development of the frontend the ellipsis, i.e. ..., was used to signify that a field defined on any dimensions is permitted as an input to a field operator or program. The concept was never fully supported and at least partially broken. While fixing typing errors this became a burden and it was decided to remove the concept until we introduce a more powerful type system that is capable of this in a proper way.