Add BlockStmt (#1118)
During development of if-statements #1079 a new FOAST node BlockStmt was introduced. The node is used to group multiple foast.Stmts into a single node (e.g. each branch of a statement is a single BlockStmt). Primarily to avoid #1079 to rot before the conceptual problems on ITIR are solved these changes were factored out into this PR. However the changes are meaningful on their own as they introduce a cleaner lowering mechanism from the somewhat-stateful FOAST to the functional ITIR.

Additional changes in this PR:

    Fixed small bug in test_closure_symbols
    Add if_ helper function to itir_makers.py