Imperative Backend (#1090)
## Description

This PR introduces a new backend that commits the expression in `gtfn_ir` to individual statements. This PR is the implementation of [this project](https://hackmd.io/@gridtools/Bk0tTITVi) of cycle 12. The purpose of (at least this iteration) of this project is to make the generated code as generated by the stencils in `icon-exclaim` more readable. This task is not well defined by definition, besides personal preference being part of the definition of readability, `itir` may represent arbitrarily complex constructs, and readability may become an ill posed requirement in those cases. Hence the focus on the patterns actually used by the `icon-exclaim` set of stencils. 

### Overview

The backend was realized as an additional lowering step with associated codegen. Optionally, `gtfn_ir` is lowered further to `gtn_im_ir` (`im` for imperative) that additionally features statements. A very lightweight codegen step is introduced transform these new dialect nodes to c++ code (`gtfn`). This can be controlled via an additional command line flag (communicated through the pipeline using `kwargs`). 

### Limitations

* The gtfn to gtfn_im lowering makes no efforts to lower constructs that pass lambdas as arguments to anything more readable
* The cse pass was instructed to not allow collection of reduction nodes. This is to ensure that reductions can be lowered to something more readable (c.f. limitation above), and to avoid situations where the call of the reduction is split from it's arguments (reductions currently need to have at least one partially shifted field as argument to the node directly).

### Testing
 
* The backend has been added to the set of test backends and disabled/enabled as consitent with the default `gtfn` backend
* The backend is tested using the full set of `icon-exclaim` stencils