Functional: Parametrize Iterators With Location Types in Type Inference (#1037)
Adds two type parameters to iterator types: the current location type (that is, where the iterator is currently located) and the defined location (that is, the location type where the iterator can be dereferenced).

In the current implementation, integer arguments passed to shift are ignored. That is, partial shifts are not tracked (and checked) by the type system.