Fix typing of `cast_` builtin in ITIR type inference (#1182)
The typing inference rule for cast_ was not correctly implemented when the builtin was introduced. This PR fixes that by making cast_ part of the grammar (similar to how this is handled in the frontend). This restriction could be lifted in the future (e.g. by introducting a type of a type).