Extend ITIR type inference to collect the types of all subexpressions (#1155)
Type inferrence on ITIR was previously only returning the type of a single expression. This PR introduces extends the inference to collect the type of all subexpressions and introduces a new function _infer_all returning a mapping from the (Python) id of a node to its type. As we start using this function we will make it public.