feat[next]: DaCe scan operator (#1284)
This commit adds support for the scan operator in DaCe backend. It implements the translation capability from iterator IR to DaCe SDFG when the stencil node contains a scan closure. The scan is translated into a loop over the vertical dimension (column axis), within a map closure in the horizontal domain.