refactor[next]: use __gt_dims__ in LocatedField (#1307)
Rename `LocatedField.axes` to `LocatedField.__gt_dims__` to make it compatible with the Cartesian and as preparation to switch to the new field implementation.

Note that gt4py.cartesian requires returning a tuple of `str`, for gt4py.next we require `tuple[Dimension]`. I propose to relax the cartesian requirement to convertible to `str`, which should already work as the documentation demands to go through `gt4py.storage.utils.get_dims()` which has string conversion.