feature[next]: Add argument types to ITIR when tracing (#1318)
Enhances the ITIR fencil tracer by populating the `dtype`, `kind` attributes of the fencil parameters (i.e. `itir.Sym`s) for all arguments that are fields. This is required by the Temporary pass in #1271. Since the ITIR type system is to be refactored only the minimal functionality needed is implemented here.