fix[next]: Fix for GPU test execution (#1368)
This commit removes test_gpu_backend.py

Spack build of icon4py was broken because dace is an optional
module, not installed in the default environment, and the dace
backend is not available in test execution. This caused
an ImportError exception in test_gpu_backend.py, because this
test is bypassing the test exclusion matrix.
The initial proposed fix was to use try/except to handle this case.
However, all tests in baseline are already executed on the GPU
backends (both GTFN and DaCe), therefore this simple test is no
longer needed.