Added a function to get the arguments to call an SDFG.
This commit adds a function that allows to generate the arguments needed to call an SDFG, before this was part of `run_dace_iterator()`.
This made it very complex to run an SDFG outside this function.

One should consider this as an amend to [PR #1379](https://github.com/GridTools/gt4py/pull/1379).