Updated how the signature of the SDFG is generated.
Before the argument list, i.e. the arguments that could be passed as positional argument, include all arguments of the SDFG.
However, due to several reasons it was changed to only include the arguments that where known to the Fencil as positional arguments.
The other arguments, such as offset providers and the ones generated by the translation process, are only accessible as keywords only arguments.