Updated again the computation/extraction of the SDFG arguments.
In commit d9164978 I thought it was necessary to use `SDFG.signature_arglist()` instead of `SDFG.arg_names`, as it turned out it was enough to give the SDFG the correct argument list.
I also updated the composition of `expected_args`, now it will be an error if an argument is missing.
It now works.