fix[next][dace]: Fix for neighbor reduction with skip values (#1443)
This PR provides a bugfix for the case of neighbor reductions with lambda function as reduction operation and connectivity table containing skip values.
The lambda function should only accumulate the results for the valid neighbors. On the contrary, the baseline implementation was using the reduction identity value for the missing neighbors, resulting in invalid result.
The fix consists of producing an array of boolean flags to determine if the neighbor value is valid or not. If not valid, the call to the lambda function is by-passed.