fix[next][dace]: Fix translation of if statement from tasklet to inter-state condition (#1469)
The bug addressed by this PR is that if-nodes were translated to tasklets. Tasklets assume that all inputs are evaluated. For if-nodes, we need to enforce exclusive execution of one of the two branches. That means that only one of the two arguments will be evaluated at runtime. We achieve this by implementing the true/false branches as separate states and checking the if-statement as condition on the inter-state edge.