feat[next][dace]: make canonical representation of field domain optional (#1476)
Baseline implementation of DaCe backend was reordering the dimensions in
field domain based on alphabetical order. This is the canonical
representation of field domain, and provides the advantage of not
requiring regenerating the SDFG for different memory layouts of field
arguments. Besides, the code for accessing a field is simple, because
all field domains are assumed to follow the same layout.

However, the canonical representation poses an obstacle to the
realization of module-level SDFGs, because it requires an additional
conversion step of all array arguments before calling the SDFG.
Therefore, we make the canonical representation optional. Note that this
change should not have any performance impact, because the real memory
layout of field arrays is not modified.