bug[next]: Mark cupy tests as requires gpu (#1483)
It noticed we execute some tests that use cupy in the cpu CSCS-CI. These tests are always executed when cupy is installed (which usually means we have a gpu, but not necessarily). This PR changes this such that they are marked with `requires_gpu` and as such can be disabled.