refactor[next]: Fencil to itir.Program for gtfn (#1524)
First PR for preparing itir for a combined field view + iterator
representation.

Adds new nodes:
- itir.Program (to replace itir.FencilDefinition)
- itir.SetAt (to replace itir.StencilClosure)
and a new builtin `as_fieldop`.

The semantic of `SetAt` is that the `expr` is directly computed into the
`target` field. `as_fieldop` aka map, takes an itir stencil (a function
from iterators to values) and promotes it to a field_operator, i.e. a
function from fields to fields.

The idea for `itir.Program` is to align and ultimately merge with
`past.Program`.

In this PR, the transition from Fencil to Program is implemented only
for GTFN with a separate pass before lowering to gtfn.

Additionally the pretty_printer/parser is extended with the new
nodes/builtin.

---------

Co-authored-by: Till Ehrengruber <t.ehrengruber@me.com>