refactor[next]: move lift_mode itir test fixture into program_processor (#1533)
Currently we have a mix of specifying the backend which already comes
with a lift_mode default and separately the lift_mode fixture in some
tests. The default was not overwritten in the roundtrip backend. Now we
remove the separate `lift_mode` and add extra backends with the
lift_mode set.

Note: we don't run double_roundtrip with temporaries.

Longer term we should refactor all itir tests to use the ffront test
infrastructure.

---------

Co-authored-by: Till Ehrengruber <t.ehrengruber@me.com>