feat[cartesian]: gtc `cuda` backend deprecation (#1498)
## Description

GTC `cuda` backend was made available a few years ago for AI2 team
research. It has been kept updated but a recent poll shows that it is
not in use. Recent new features break the backend and we propose here to
hard deprecate it rather than keep spending time maintaining it.

`GT4PY_GTC_ENABLE_CUDA=1` can be used to force the use of the backend, but
will warn that any feature from February 2024 are not available/not
tested.

Additionally a mechanism to deprecate all GTC backends are now in use. Using
```python
@disabled(
    message="Disable message.",
    enabled_env_var="EnvVarToEnable",
)
```

## Requirements

- [x] All fixes and/or new features come with corresponding tests.

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>