Updated and fixed a big in the `is_interstate_transient()` function.
There was a bug before, as the set was computed wrongly, it sued the sink instead of the source nodes.
However, I have updated and renamed the function, it now works differently, i.e. not if we can remove it, but if we have to keep it.
Also after the discussion with Hannes and Edoardo I refined the definition of the set.