feat[next][dace]: Skeleton of GTIR DaCe backend (#1538)
We introduce a new DaCe backend that translates a GTIR program to SDFG.
GTIR is still sometimes referred to as Combined-IR and provides a
fieldview representation of the GT4Py program. The GT4Py frontend is not
ready to lower a GT4Py program to GTIR, so for now the unit tests use
manually written GTIR programs as input.

The new DaCe backend is implemented in the package
`gt4py.next.program_processors.runners.dace_fieldview`
This package provides, for now, the method `build_sdfg_from_gtir` to
build the DaCe SDFG.

This PR introduces translation capability for the following GTIR
types:`ir.Program`, `ir.Stmt`, `ir.SetAt`. And for the following new
builtin functions: `as_field_op` and `cond`.