feat[next][dace]: GTIR-to-SDFG lowering of shifts (#1577)
This PR adds support for lowering of the `shift` operator to DaCe SDFG.
It supports both cartesian and unstructured shift. The shift also works
with dynamic offset. The lowering is designed in a way to support pure
fieldview IR as well as mixed IR with composition of shift operators.

Additional changes:
- type inference is enabled
- the IR maker `as_fieldop` is imported from `gtir` branch and the IRs
in dace tests are updated to use it
- the storage for literal values and symbols is changed to scalar (was
array before)
- removed extra call around the `cond` primitive, which was not needed
since no arguments are passed