feat[next][dace]: GTIR-to-SDFG lowering of let-lambdas (using NestedSDFG) (#1601)
This PR proposes an alternative design for the lowering of let-lambdas
to SDFG, than the design provided in #1589. The previous PR added a
separate state for the definition of let-symbols: the disadvantage of
that solution is that the generated SDFG does not clearly represent the
data dependency. The alternative design proposed in this PR uses a
nested SDFG to represent some local symbols in the lambda scope. The
lambda function is lowered inside the nested SDFG. We rely on the
simplify pass to remove unnecessary nested SDFGs.