refactor: Remove stale references to flake8 (#1625)
In PR https://github.com/GridTools/gt4py/pull/1466, `flake8`, `black`,
and `isort` were replaced with the [ruff](https://astral.sh/ruff) linter
and formatter. It seems there are a few leftovers of this transition. In
this PR, I did a quick search for flake8 and deleted / replaced all
instances found.

Removing `flake8` means rebuilding the requirements (to propagate the
changes to the frozen `requirements-*` files), which updates a couple of
dependencies.

We also decided to remove the `linter-*` tasks of `tox` since they have
been replaced by `pre-commit`.