refactor[cartesian]: Enable MaskInlining in gt4yp - DaCe bridge (#1624)
`MaskInlining` will be a key feature of the improved gt4py / DaCe bridge
as envisioned in https://github.com/GEOS-ESM/NDSL/issues/53, which will
propagate more control flow elements from gt4py to DaCe. We thus want to
make sure if `MaskInlining` itself is posing any problems. We tracked
the git history of this explicit exclude back to PR
https://github.com/GridTools/gt4py/pull/300, which not only added the
npir backend, but also refactored how pipeline stages are defined.
Previous to this PR optimization passes to be specified explicitly and
the PR changed this to an exclude list. It looks like the `MaskInlining`
pass was added after the list of optimization passes was formed and thus
excluded (to keep the list constant). From this history we thus don't
expect `MaskInlining` to break anything.

Co-authored-by: Roman Cattaneo <>