refactor[cartesian]: readability improvements in gtir -> oir conversion and other cleanups (#1630)
Being new to the codebase and in preparation of the [GT4Py-DaCe bridge
refactor](https://github.com/GEOS-ESM/NDSL/issues/53) I was reading a
lot of code. Some things that caught my eye are summarized in this PR.
Bigger changes include

1. Moving `daceir.py` into the `dace/` backend folder. This follows the
convention of all other backends who have their IR inside the backend
folder. (last commit)
2. Readability improvements in `gtir_to_oir.py` (the "Cleanup visit_*"
commits): In my opinion, there's no point in shortening "statement" to
"stmt". I wouldn't go as far as renaming the classes. For the renamed
local variables, I think this adds a lot to readability.
3. `visit_While` in `gtir_to_oir.py` was having an extra mask statement
around the `while` loop (in case a mask was defined. I inlined the
potential `mask` with the condition of `while` loop.
4. In `gtir_to_oir.py` don't translate every body statement into a
single `MaskStmt`. Instead, create one (or two incase of `else`)
`MaskStmt` with multiple statements in the `body` (which is already
typed as a list of statements).

Tested locally by running the test suite. All tests passing.

Co-authored-by: Roman Cattaneo <>