bug[next]: fix to gtir lowering of `astype` on tuples containing scalars (#1642)
The unpacked tuple will contain `as_fieldop`ed `cast_`s for fields, but
direct `cast_`s for scalars.

Note: A similar change is not needed for `where` which also works on
tuples in the `true` and `false` branches. The reason is that `mask` is
required to be a `Field`, therefore `true` and `false` are promoted to
`Field`s, too.